+++ packages/web/src/__stories__/componentConfigStickerSheet/StickerSheet.tsx
@@ -11,7 +11,6 @@ import { Chip } from '@coinbase/cds-web/chips/Chip';
 import { InputChip } from '@coinbase/cds-web/chips/InputChip';
 import { MediaChip } from '@coinbase/cds-web/chips/MediaChip';
 import { Coachmark } from '@coinbase/cds-web/coachmark/Coachmark';
-import { DotCount } from '@coinbase/cds-web/dots/DotCount';
 import { Icon } from '@coinbase/cds-web/icons/Icon';
 import { Pictogram } from '@coinbase/cds-web/illustrations/Pictogram';
 import { HStack } from '@coinbase/cds-web/layout/HStack';
@@ -25,6 +24,7 @@ import { Text } from '@coinbase/cds-web/typography/Text';
 import { AlertExample } from './examples/AlertExample';
 import { ControlsExample } from './examples/Controls';
 import { DatePickerExample } from './examples/DatePicker';
+import { DotCountExample } from './examples/DotCount';
 import { DropdownExample } from './examples/DropdownExample';
 import { ModalExample } from './examples/ModalExample';
 import { PaginationExample } from './examples/Pagination';
@@ -397,15 +397,7 @@ export const StickerSheet = memo(() => {
           </Container>
           <VStack className="no-a11y-checks">
             <Container title="DotCount">
-              <DotCount count={3}>
-                <Icon name="bell" size="l" />
-              </DotCount>
-              <DotCount count={12}>
-                <Icon name="bell" size="l" />
-              </DotCount>
-              <DotCount count={100} max={99}>
-                <Icon name="bell" size="l" />
-              </DotCount>
+              <DotCountExample />
             </Container>
           </VStack>
         </VStack>