+++ packages/mcp-server/CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 <!-- template-start -->
 
+## 9.1.0 ((5/26/2026, 01:55 PM PST))
+
+This is an artificial version bump with no new change.
+
 ## 9.0.2 ((5/22/2026, 09:54 AM PST))
 
 This is an artificial version bump with no new change.